#8 feat: improved node serialization and fixed related bugs
- using kademlia-api 4.1.13

+++ src/main/java/io/ep2p/kademlia/netty/NettyKademliaDHTNode.java
@@ -9,16 +9,15 @@ import lombok.Getter;
 import lombok.SneakyThrows;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.concurrent.Future;
 
 public class NettyKademliaDHTNode<K extends Serializable, V extends Serializable>
-        extends DHTKademliaNodeAPIDecorator<BigInteger, NettyConnectionInfo, K, V> {
+        extends DHTKademliaNodeAPIDecorator<Long, NettyConnectionInfo, K, V> {
 
     @Getter
     private final transient KademliaNodeServer<K, V> kademliaNodeServer;
 
-    public NettyKademliaDHTNode(DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> kademliaNode, KademliaNodeServer<K, V> kademliaNodeServer) {
+    public NettyKademliaDHTNode(DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> kademliaNode, KademliaNodeServer<K, V> kademliaNodeServer) {
         super(kademliaNode);
         this.kademliaNodeServer = kademliaNodeServer;
     }
@@ -32,7 +31,7 @@ public class NettyKademliaDHTNode<K extends Serializable, V extends Serializable
 
     @Override
     @SneakyThrows
-    public Future<Boolean> start(Node<BigInteger, NettyConnectionInfo> bootstrapNode) {
+    public Future<Boolean> start(Node<Long, NettyConnectionInfo> bootstrapNode) {
         kademliaNodeServer.run(this);
         return super.start(bootstrapNode);
     }
