+++ src/main/java/io/ep2p/kademlia/netty/serialization/NodeDeserializer.java
@@ -6,15 +6,15 @@ import io.ep2p.kademlia.netty.common.NettyExternalNode;
 import io.ep2p.kademlia.node.Node;
 
 import java.lang.reflect.Type;
-import java.math.BigInteger;
 
-public class NodeDeserializer implements JsonDeserializer<Node<BigInteger, NettyConnectionInfo>> {
+
+public class NodeDeserializer implements JsonDeserializer<Node<Long, NettyConnectionInfo>> {
     @Override
-    public Node<BigInteger, NettyConnectionInfo> deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
+    public Node<Long, NettyConnectionInfo> deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
         NettyExternalNode nettyExternalNode = new NettyExternalNode();
         JsonObject jsonObject = jsonElement.getAsJsonObject();
 
-        nettyExternalNode.setId(jsonDeserializationContext.deserialize(jsonObject.get("id"), BigInteger.class));
+        nettyExternalNode.setId(jsonDeserializationContext.deserialize(jsonObject.get("id"), Long.class));
         nettyExternalNode.setConnectionInfo(jsonDeserializationContext.deserialize(jsonObject.get("connectionInfo"), NettyConnectionInfo.class));
         return nettyExternalNode;
     }
