+++ src/test/java/io/ep2p/kademlia/netty/Example.java
@@ -11,7 +11,6 @@ import io.ep2p.kademlia.node.KeyHashGenerator;
 import io.ep2p.kademlia.util.BoundedHashUtil;
 import lombok.SneakyThrows;
 
-import java.math.BigInteger;
 import java.util.concurrent.TimeUnit;
 
 public class Example {
@@ -24,18 +23,18 @@ public class Example {
 
         OkHttpMessageSender<String, String> okHttpMessageSender = new OkHttpMessageSender<>();
 
-        KeyHashGenerator<BigInteger, String> keyHashGenerator = key -> {
+        KeyHashGenerator<Long, String> keyHashGenerator = key -> {
             try {
-                return new BoundedHashUtil(NodeSettings.Default.IDENTIFIER_SIZE).hash(key.hashCode(), BigInteger.class);
+                return new BoundedHashUtil(NodeSettings.Default.IDENTIFIER_SIZE).hash(key.hashCode(), Long.class);
             } catch (UnsupportedBoundingException e) {
                 e.printStackTrace();
             }
-            return BigInteger.valueOf(key.hashCode());
+            return Long.valueOf(key.hashCode());
         };
 
 
         NettyKademliaDHTNode<String, String> node1 = new NettyKademliaDHTNodeBuilder<>(
-                BigInteger.valueOf(2),
+                2L,
                 new NettyConnectionInfo("127.0.0.1", 8000),
                 new SampleRepository(),
                 keyHashGenerator
@@ -45,7 +44,7 @@ public class Example {
 
         // node 2
         NettyKademliaDHTNode< String, String> node2 = new NettyKademliaDHTNodeBuilder<>(
-                BigInteger.valueOf(2),
+                2L,
                 new NettyConnectionInfo("127.0.0.1", 8001),
                 new SampleRepository(),
                 keyHashGenerator
@@ -53,11 +52,11 @@ public class Example {
 
         System.out.println("Bootstrapped? " + node2.start(node1).get(5, TimeUnit.SECONDS));
 
-        StoreAnswer<BigInteger, String> storeAnswer = node2.store("K", "V").get();
+        StoreAnswer<Long, String> storeAnswer = node2.store("K", "V").get();
         System.out.println(storeAnswer.getResult());
         System.out.println(storeAnswer.getNodeId());
 
-        LookupAnswer<BigInteger, String, String> k = node1.lookup("K").get();
+        LookupAnswer<Long, String, String> k = node1.lookup("K").get();
         System.out.println(k.getResult());
         System.out.println(k.getValue());
